Change Session.withScale to a double and adjust Driver tests

+++ src/main/java/com/teradata/tpcds/Session.java
@@ -82,7 +82,7 @@ public class Session
         );
     }
 
-    public Session withScale(float scale)
+    public Session withScale(double scale)
     {
         return new Session(
                 scale,
